Move render_macro_matcher to own module
